Telos 2.2.2: Detected another bug that only stores integer values instead of float values for the points. Also, once the task is cancelled, the ending date should be inserted.

+++ task/models.py
@@ -48,7 +48,7 @@ class Task(models.Model):
     category = models.CharField(
         max_length=24, choices=CATEGORIES, default=PERSONAL_DEVELOPMENT)
     status = models.CharField(max_length=24, choices=STATUS, default=ACTIVE)
-    points = models.IntegerField(default=5)
+    points = models.FloatField(default=5)
     life_task = models.IntegerField(default=3) # task have a life of 4 weeks to be completed
 
     initial_week = models.CharField(max_length=2, null=False, default=date.today().isocalendar()[1])
